core: overhaul runtime state and parser logic
Refactored the entire evaluation flow to be more explicit. Removed
the unnecessary abstraction layers in `ppc.c` and `eval.c`.
The parser is now leaner, focusing on direct evaluation without
the overhead of managing multiple run modes.

Summary of changes:
- Dropped `instr/home.c`, `instr/list.c`, `instr/run.c`.
- Tightened `PPC_Runtime` and `PPC_Ctx` structs.
- Optimized line parsing and memory initialization.

+++ instr/trans.c
@@ -11,7 +11,7 @@ MKINSTR(trans)
 {
 	_ARGC_MIN(2)
 
-	int character = ctx->argv[1][0];
+	uint8_t character = ctx->argv[1][0];
 	IFNPIPE printf("%d\n", character);
 	return VAL_INT(character);
 }
